# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GTG API Documentation",
        "description": "API Documentation for GTG Application",
        "contact": {
            "name": "GTG Support",
            "email": "support@gtg.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/fetch-accounts": {
            "post": {
                "tags": [
                    "Fetch Accounts"
                ],
                "summary": "Get all accounts for a user",
                "description": "Retrieves all account numbers associated with the given user ID",
                "operationId": "getAllAccounts",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "userID"
                                ],
                                "properties": {
                                    "userID": {
                                        "type": "string",
                                        "example": "Testi12059528"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Accounts retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "accounts": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "account_no": {
                                                                "type": "string",
                                                                "example": "1234567890"
                                                            },
                                                            "account_type": {
                                                                "type": "string",
                                                                "example": "savings"
                                                            },
                                                            "account_status": {
                                                                "type": "string",
                                                                "example": "active"
                                                            },
                                                            "opening_date": {
                                                                "type": "string",
                                                                "format": "date-time"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The userID field is required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "An error occurred while fetching accounts"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "X-API-Key": []
                    },
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/fetch-account-name": {
            "post": {
                "tags": [
                    "Get Account Name"
                ],
                "summary": "Get user's full name by userID",
                "description": "Retrieves the user's full name based on their userID",
                "operationId": "getAccountName",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "userID"
                                ],
                                "properties": {
                                    "userID": {
                                        "description": "User ID to lookup",
                                        "type": "string",
                                        "example": "Webst446399507"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Account name retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "full_name": {
                                                    "type": "string",
                                                    "example": "John Smith Doe"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The userID field is required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "X-API-Key": []
                    },
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/account/balance": {
            "post": {
                "tags": [
                    "Account Balance"
                ],
                "summary": "Get account balance",
                "description": "Retrieves the current balance and transaction totals for an account",
                "operationId": "getAccountBalance",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "AccountNo"
                                ],
                                "properties": {
                                    "AccountNo": {
                                        "type": "string",
                                        "example": "1234567890"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Request successfully completed"
                                        },
                                        "data": {
                                            "properties": {
                                                "balance": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": "-200"
                                                },
                                                "credit_total": {
                                                    "type": "string",
                                                    "example": "138000.0000000000"
                                                },
                                                "debit_total": {
                                                    "type": "string",
                                                    "example": "138200.0000000000"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "response_time": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 1949.21
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Account number is required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "X-API-Key": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/account/name": {
            "get": {
                "tags": [
                    "Account"
                ],
                "summary": "Get account holder name by account number",
                "operationId": "1d0e74b13bc81f177c7f87e38e0feeee",
                "parameters": [
                    {
                        "name": "account_no",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Account not found"
                    }
                }
            }
        },
        "/api/v1/account/type": {
            "post": {
                "tags": [
                    "Account"
                ],
                "summary": "Fetch account type",
                "operationId": "73948a857516b5bbda41b6db051e32c7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FetchAccountTypeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/v1/account/status": {
            "post": {
                "tags": [
                    "Account"
                ],
                "summary": "Fetch account status by account number",
                "operationId": "0098333ee196a3285c4a14b8ead58918",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FetchAccountStatusRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/v1/account/number": {
            "post": {
                "tags": [
                    "Account"
                ],
                "summary": "Get account number by user credentials",
                "operationId": "4a84732bde9faac7a6de36dfb2321dbb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FetchAccountNumberRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/v1/passport/upload": {
            "post": {
                "tags": [
                    "Passport"
                ],
                "summary": "Upload user passport",
                "operationId": "cf204e764e8c4a260181c7a67fef835d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/UploadPassportRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "FetchAccountTypeRequest": {
                "required": [
                    "account_no"
                ],
                "properties": {
                    "account_no": {
                        "type": "string",
                        "example": "1000000001"
                    }
                },
                "type": "object"
            },
            "FetchAccountStatusRequest": {
                "required": [
                    "account_no"
                ],
                "properties": {
                    "account_no": {
                        "type": "string",
                        "example": "1000000001"
                    }
                },
                "type": "object"
            },
            "FetchAccountNumberRequest": {
                "required": [
                    "user_id",
                    "password"
                ],
                "properties": {
                    "user_id": {
                        "type": "string",
                        "example": "Testi12059528"
                    },
                    "password": {
                        "type": "string",
                        "example": "your_password"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "X-API-Key": {
                "type": "apiKey",
                "name": "X-API-Key",
                "in": "header"
            },
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}